adding further instructions in docs
