--- data_generator/main_pipeline_cli.py
@@ -1,380 +1,0 @@
-'''
-Main script to run the validation tool
-Author: Pearl Park
-required files: /data_generator/dg_config.json /config/openai-key.ini
-'''
-import os, shutil, sys
-import configparser
-import json 
-from icecream import ic
-from datetime import datetime
-import argparse
-
-import copy
-import click
-if os.environ.get('INDALEKO_ROOT') is None:
-    current_path = os.path.dirname(os.path.abspath(__file__))
-    while not os.path.exists(os.path.join(current_path, 'Indaleko.py')):
-        current_path = os.path.dirname(current_path)
-    os.environ['INDALEKO_ROOT'] = current_path
-    sys.path.append(current_path)
-    
-from data_generator.scripts.s1_metadata_generator import Dataset_Generator
-from data_generator.scripts.s2_store_test_Indaleko import MetadataStorer
-from data_generator.scripts.s3_translate_query import QueryExtractor
-from data_generator.scripts.s4_translate_AQL import AQLQueryConverter
-from data_generator.scripts.s5_get_precision_and_recall import ResultCalculator
-from db.db_config import IndalekoDBConfig
-from db.i_collections import IndalekoCollections
-from query import NLParser, QueryHistory, AQLExecutor, OpenAIConnector
-from data_generator.scripts.s6_log_result import ResultLogger
-from pathlib import Path
-from typing import Dict, Any
-import subprocess
-from db.db_collection_metadata import IndalekoDBCollectionsMetadata
-
-class Validator():
-    """
-    This is the class for performing the validation for Indaleko
-    """
-    def __init__(self, args) -> None:
-        self.file_path = "./data_generator/results/"
-        self.stored_file_path = self.file_path + "stored_metadata/"
-        self.config_path =  './data_generator/config/'
-
-        api_path = './config/openai-key.ini'
-
-        path = Path(self.file_path)
-        if path.exists():
-            shutil.rmtree(path)
-        path.mkdir(parents=True, exist_ok=True)
-
-        metadata_path = Path(self.stored_file_path)
-        metadata_path.mkdir(parents=True, exist_ok=True)
-
-        self.config = self.get_config_file(self.config_path + 'dg_config.json')
-        self.logger = ResultLogger(result_path=self.file_path)
-        self.db_config = IndalekoDBConfig()
-        self.db_config.setup_database(self.db_config.config['database']['database'])
-
-        # if not args.no_reset:
-        #     try:
-        #         self.db_config.reset_command()
-        #         # subprocess.run(["python3", "./db/db_config.py", "reset"], check=True)
-        #         subprocess.run(["python3", "./storage/recorders/local/mac/recorder.py", "--arangoimport"], check=True)
-        #         subprocess.run(["python3", "./platforms/mac/machine_config.py", "--add"], check=True)
-        #     except subprocess.CalledProcessError as e:
-        #         raise e
-
-        self.db_config.collections = IndalekoCollections()
-
-        self.query_extractor = QueryExtractor()
-        self.data_generator = Dataset_Generator(self.config)
-        self.data_storer = MetadataStorer()
-        self.result_calculator = ResultCalculator()
-
-        self.openai_key = self.get_api_key(api_path)
-        self.llm_connector = OpenAIConnector(api_key=self.openai_key)
-        self.collections_md = IndalekoDBCollectionsMetadata(self.db_config)
-        self.nl_parser = NLParser(self.llm_connector, self.collections_md)
-        self.query_translator = AQLQueryConverter()
-        self.query_executor = AQLExecutor()
-
-        self.dynamic_activity_providers = {}
-        self.result_dictionary = {}
-        self.schema = self.read_json('data_generator/config/schema.json')
-    
-    def time_operation(self, operation, **kwargs) -> tuple[str, Any]:
-        """
-        Given a function, return the time and results of the operation
-        Args:
-            operation (Any): The function or operation to time
-            **kwargs (Any): The arguments required to run the function
-        Returns:
-            tuple[str, Any]: The operation time and any outputs returned 
-        """
-        start_time = datetime.now()
-        results = operation(**kwargs)
-        end_time = datetime.now()
-        operation_time = end_time - start_time
-        return str(operation_time), results
-
-    def get_api_key(self, api_key_file: str) -> str:
-        """
-        Given the path to an API key file in the config dir, return the api key
-        Args:
-            api_key_file (str): The path to the api key file
-        Returns:
-            str: The API key
-        """
-        assert os.path.exists(api_key_file), \
-            f"API key file ({api_key_file}) not found"
-        config = configparser.ConfigParser()
-        config.read(api_key_file, encoding='utf-8-sig')
-        openai_key = config['openai']['api_key']
-        if openai_key is None:
-            raise ValueError("OpenAI API key not found in config file")
-        if openai_key[0] == '"' or openai_key[0] == "'":
-            openai_key = openai_key[1:]
-        if openai_key[-1] == '"' or openai_key[-1] == "'":
-            openai_key = openai_key[:-1]
-        return openai_key
-
-    def get_config_file(self, config_path) -> dict[str]:
-        """
-        Given the path to a config file, parse the configuration file in the form of a dictionary
-        Args:
-            config_path (str): The path to the config file
-        Returns:
-            Dict[str]: The config file
-        """
-        assert os.path.exists(config_path), f'Config file path "{config_path}" not found'
-        with open(config_path, 'r') as file:
-            config = json.load(file)
-        return config
-    
-    def add_result_to_dict(self, key: str, value: str) -> None:
-        """
-        Adds the results to the result dictionary
-        Args:
-            key (str): The key to add to the dictionary
-            value (str): The value to add to the dictionary
-        """
-        self.result_dictionary[key] = value
-    
-    def generate_query(self, query: str, n_truth_md: int, query_attributes) -> str:
-        # PARSE QUERY
-        # Adapted from IndalekoSearch.py
-        # NOTE: parser hasn't yet been implemented so takes the query + schema and puts in dictionary format
-        parse_query_time, parsed_query = self.time_operation(
-            self.nl_parser.parse, 
-            query=query)
-
-        self.logger.log_process("parsing query...")
-        ic(f"Parse time: {parse_query_time}")
-        self.logger.log_process_result("parsed_query", parse_query_time) 
-
-        #GENERATE AQL TRANSLATION:
-        translate_query_time, translated_query = self.time_operation(
-            self.query_translator.translate, 
-            parsed_query=parsed_query, 
-            selected_md_attributes=query_attributes["converted_selected_md_attributes"], 
-            dynamic_activity_providers=self.dynamic_activity_providers,
-            additional_notes=query_attributes["geo_coords"], 
-            n_truth = n_truth_md,
-            llm_connector=self.llm_connector)
-        self.logger.log_process_result("translated_aql", translate_query_time, translated_query)
-        return translated_query
-
-    def generate_dictionary(self, query:str):
-        json_name = "dictionary.json"
-        self.logger.log_process("building dictionary...")
-        selected_md_attributes = self.query_extractor.extract(query = query, llm_connector = self.llm_connector)
-        self.logger.log_process_result("selected_md_attributes", selected_md_attributes)
-        self.write_as_json(self.config_path, json_name, selected_md_attributes)
-        dictionary_selection = click.prompt("Dictionary ready for evaluation, please type 1 to continue, otherwise type 0 to exit", type=int)
-        if dictionary_selection != 1:
-            sys.exit()
-        selected_md_attributes = self.read_json(self.config_path + json_name)    
-        return selected_md_attributes
-
-    def run(self) -> None:
-        """
-        Run function for the validator tool
-        """
-        translated_query ={}
-        query = self.config["query"]
-        ic(query)
-        n_truth_md = self.config["n_matching_queries"]
-        n_total_md = self.config["n_metadata_records"]
-
-        self.logger.log_config(self.config)
-        self.add_result_to_dict("query", query)
-        self.add_result_to_dict("n_total_truth", n_truth_md)
-        self.add_result_to_dict("n_metadata", n_total_md)
-
-        intial_selection = click.prompt("Create new dictionary (1) or use existing (0)", type=int)
-        if intial_selection == 1:
-            print("Creating new dictionary...")
-            selected_md_attributes = self.generate_dictionary(query)
-        elif intial_selection == 0:
-            print("Using existing dictionary...")
-            selected_md_attributes = self.read_json(self.config_path + "dictionary.json")  
-        else:
-            print("Invalid input. Exiting.")
-            sys.exit()
-
-        self.add_result_to_dict("selected_md_attributes", selected_md_attributes)
-        #selected_md_attributes = self.data_generator.preprocess_dictionary_timestamps(selected_md_attributes, False)
-
-        # GENERATE METADATA DATASET:        
-        self.logger.log_process("generating record and activity metadata...")
-        copy_selected_md = copy.deepcopy(selected_md_attributes)
-        generation_time, results = self.time_operation(self.data_generator.generate_metadata_dataset, 
-                                                        selected_md_attributes=copy_selected_md)
-        (
-            all_records_md, 
-            all_geo_activity_md, 
-            all_temp_activity_md, 
-            all_music_activity_md, 
-            all_machine_config_md, 
-            all_semantics_md, 
-            stats
-        ) = results
-
-        # save the resulting dataset to a json file for future reference
-        self.data_generator.write_json(all_records_md, self.stored_file_path + "records.json")
-        self.data_generator.write_json(all_geo_activity_md, self.stored_file_path + "geo_activity.json")
-        self.data_generator.write_json(all_music_activity_md, self.stored_file_path + "music_activity.json")
-        self.data_generator.write_json(all_temp_activity_md, self.stored_file_path + "temp_activity.json")
-        self.data_generator.write_json(all_machine_config_md, self.stored_file_path + "machine_config.json")
-        self.data_generator.write_json(all_semantics_md, self.stored_file_path + "semantics.json")
-
-        self.logger.log_process_result("metadata_generation_time", generation_time)
-        self.add_result_to_dict("metadata_stats", stats)
-        
-        # -------------------------------------------------------------------------------------
-        # STORE METADATA INTO DATABASE:
-        self.logger.log_process("storing record metadata...")
-        record_storage_time = self.time_operation(self.data_storer.add_records_to_collection, 
-                                                    collections=self.db_config.collections, 
-                                                    collection_name="Objects", 
-                                                    records=all_records_md)
-        ic(f"Storing time for record metadata: {record_storage_time}")
-        self.logger.log_process_result("stored record:", record_storage_time[0])
-       
-        geo_registrator, geo_collection = self.data_storer.register_activity_provider("Geographical Location Collector")
-        self.dynamic_activity_providers["GeoActivity"]=geo_registrator.get_activity_collection_name()
-
-        geo_activity_storage_time = self.time_operation(self.data_storer.add_records_with_activity_provider, collection=geo_collection, activity_contexts=all_geo_activity_md)
-        self.logger.log_process("storing geo activity context...")
-        ic(f"Storing time for geographical location activity metadata: {geo_activity_storage_time}")
-        self.logger.log_process_result("stored geo activity context:", geo_activity_storage_time[0])
-
-        music_registrator, music_collection = self.data_storer.register_activity_provider("Music Collector")
-        self.dynamic_activity_providers["MusicActivity"]=music_registrator.get_activity_collection_name()
-
-        music_activity_storage_time = self.time_operation(self.data_storer.add_records_with_activity_provider, collection=music_collection, activity_contexts=all_music_activity_md)
-        self.logger.log_process("storing music activity context...")
-        ic(f"Storing time for music activity metadata: {music_activity_storage_time}")
-        self.logger.log_process_result("stored music activity context:", music_activity_storage_time[0])
-
-        temp_registrator, temp_collection = self.data_storer.register_activity_provider("Temperature Collector")
-        self.dynamic_activity_providers["TempActivity"]= temp_registrator.get_activity_collection_name()
-
-        temp_activity_storage_time = self.time_operation(self.data_storer.add_records_with_activity_provider, collection=temp_collection, activity_contexts=all_temp_activity_md)
-        self.logger.log_process("storing music activity context...")
-        ic(f"Storing time for music activity metadata: {temp_activity_storage_time}")
-        self.logger.log_process_result("stored music activity context:", temp_activity_storage_time[0])
-    
-        self.logger.log_process("storing semantics metadata...")
-        semantics_storage_time = self.time_operation(self.data_storer.add_records_to_collection, collections=self.db_config.collections, collection_name="SemanticData", records=all_semantics_md, key_required = True)
-        ic(f"Storing time for semantics metadata: {semantics_storage_time}")
-        self.logger.log_process_result("stored semantics metadata:", semantics_storage_time[0])
-
-        self.logger.log_process("storing machine config metadata...")
-        machine_config_storage_time = self.time_operation(self.data_storer.add_records_to_collection, collections=self.db_config.collections, collection_name="MachineConfig", records=all_machine_config_md, key_required = True)
-        ic(f"Storing time for machine config metadata: {machine_config_storage_time}")
-        self.logger.log_process_result("stored machine config:", machine_config_storage_time[0])
-
-
-        collections = self.collections_md.__activity_data_provider_collection_handler()
-        ic(collections)
-        quit()
-
-        #PREPARE FOR QUERY GENERATION:
-        self.logger.log_process("preparing aql translation...")
-        # retrieve the geo coordinates for the AQL translator:
-        translated_query["geo_coords"] = str(self.data_generator.geo_activity_generator.get_saved_geolocation())
-        self.add_result_to_dict("geo_coords", translated_query["geo_coords"])
-        
-        #convert the time to posix timestamps for consistent AQL translation results:
-        self.data_generator.posix_generator.preprocess_dictionary_timestamps(True)
-        copy_selected_md["Posix"] = self.data_generator.posix_generator.selected_md
-        translated_query["converted_selected_md_attributes"] = copy_selected_md
-        ic(translated_query)
-        self.logger.log_process("translating query to aql...")
-        self.logger.log_process_result("converted_selected_md_attributes", translated_query["converted_selected_md_attributes"])
-        self.add_result_to_dict("converted_selected_md_attributes", translated_query["converted_selected_md_attributes"])
-
-        # GENERATE QUERY:
-        query_info = "query_info.json"
-        aql_text = "AQL_query.aql"
-
-
-        translated_query["providers"] = self.dynamic_activity_providers
-        translate_query = click.prompt("Ready for AQL generation, please type 1 to use existing AQL, otherwise type 0 to generate a new one", type=int)
-        if translate_query == 1:
-            aql = self.read_aql(self.config_path + aql_text)
-        elif translate_query == 0:
-            aql = self.generate_query(query, n_truth_md, translated_query)
-        
-        self.write_as_json(self.config_path, query_info, translated_query)
-        self.write_as_aql(self.config_path, aql_text, aql)
-
-        aql_selection = click.prompt("AQL query ready for evaluation. Please review the AQL_query.aql file and query_info.json and make any necessary changes to the aql file. Type 1 to continue, otherwise type 0 to exit", type=int)
-
-        if aql_selection != 1:
-            sys.exit()
-        
-        final_query = self.read_aql(self.config_path + aql_text)
-        self.add_result_to_dict("aql_query", final_query)
-
-        # RUN INDALEKO SEARCH
-        self.logger.log_process("running search with translated aql query...")
-        execute_time, raw_results = self.time_operation(
-            self.query_executor.execute,
-            query=final_query,
-            data_connector=self.db_config
-        )
-        self.logger.log_process_result("Indaleko search completed", execute_time)
-
-        # CALCULATE PRECISION AND RECALL
-        self.logger.log_process("calculating precision and recall")
-        calculation_time, calculation_result = self.time_operation(self.result_calculator.run, raw_results=raw_results, theoretical_truth_n=n_truth_md)
-        n_actual_truth_md, precision, recall =  calculation_result[0], calculation_result[1], calculation_result[2]
-        self.logger.log_process_result("calculated precision and recall", calculation_time, f"precision: {precision}, recall: {recall}")
-        
-        self.add_result_to_dict("uuid_returned", self.result_calculator.selected_uuid)
-        self.add_result_to_dict("actual_n_total_truth", n_actual_truth_md)
-        self.add_result_to_dict("actual_n_metadata", len(raw_results))
-        self.add_result_to_dict("precision", precision)
-        self.add_result_to_dict("recall", recall)
-        self.logger.log_process("precision and recall calculated")
-        self.logger.log_process("DONE -- please check /data_generator/results/validator_result.log for the full results")
-        self.write_as_json(self.file_path, "Indaleko_search_result.json", raw_results)
-
-
-    def read_aql(self, aql_path):
-        with open(aql_path, "r") as f:
-            return f.read()
-    
-    def read_json(self, json_path):
-        with open(json_path, 'r') as file:
-            data = json.load(file)
-        return data
-
-    def write_as_aql(self, path, title, aql_query):
-        with open(path+title, "w") as file:
-            file.write(aql_query)
-
-    def write_as_json(self, path, title, result_dict):
-        with open(path + title, 'w') as file:
-            json.dump(result_dict, file, indent=4)
-
-def main() -> None:
-    '''Main function for the validator tool'''
-    parser = argparse.ArgumentParser()
-    parser.add_argument('--no_reset',
-                        '-nr',
-                        help='Prevents resetting of the database', action='store_true')
-    args = parser.parse_args()
-
-    validator_tool = Validator(args)
-    total_epoch = validator_tool.time_operation(validator_tool.run)
-    validator_tool.logger.log_final_result(total_epoch[0], validator_tool.result_dictionary)
-    
-    
-
-if __name__ == '__main__':
-    main()
